Weatherapp/requirements.txt
tk
pytz
geopy
timezonefinder
request
